Make LoginServlet properly redirect to appropriate endpoints
When a user sends their credentials via a POST request, their browser
should be redirected to the employee homepage if successful. Otherwise,
this servlet returns the login page again.

I also added a doGet method since it was required for the login endpoint
to work on browsers. If the user is already logged in, i.e. there is
already an Employee object in the session, this will immediately
redirect the browser to the employee homepage.